% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateParser.R
\name{add_rate_part_to_frame}
\alias{add_rate_part_to_frame}
\title{Evaluate a rate part and add to the dataframe.}
\usage{
add_rate_part_to_frame(df, name, name_list, class_rate, cust_class)
}
\description{
Evaluate a rate part and add to the dataframe.
}

