% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateParser.R
\name{calculate_class_bill}
\alias{calculate_class_bill}
\title{Calculate water bills for a single customer class.}
\usage{
calculate_class_bill(df, parsed_rate)
}
\arguments{
\item{df}{Data frame containing all data values referenced in the rate
file but not defined in the rate file. E.g. \code{usage_ccf},
\code{cust_class}, \code{meter_size}, \code{etc.}}

\item{parsed_rate}{\href{https://github.com/California-Data-Collaborative/Open-Water-Rate-Specification}{OWRS file},
parsed into an \code{R} list by the \code{yaml} package.}
}
\value{
Original data frame with additional columns appended. Additional columns are those
defined in the rate file, as well as columns \code{Xi} representing the usage in the
ith tier, and \code{XRi} representing revenue in the ith tier (where applicable).
}
\description{
\code{calculate_class_bill} applies a rate structure to water use data to
calculate water bills.
}

