% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateParser.R
\name{collapse_tiers}
\alias{collapse_tiers}
\title{Collapse a multi-level list.}
\usage{
collapse_tiers(map)
}
\arguments{
\item{map}{List representing a conditional/map in OWRS format.}
}
\value{
Either the original list or the collapsed list.
}
\description{
\code{collapse_tiers} collapses all elements of the 2nd level of a list into a string.
elements are pasted together separated by "\\n".
}
\keyword{internal}

