% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateParser.R
\name{eval_field_or_formula}
\alias{eval_field_or_formula}
\title{Evaluate a field or formula.}
\usage{
eval_field_or_formula(df, rate_part)
}
\arguments{
\item{df}{Data frame containing all data values referenced in field or formula.}

\item{rate_part}{List representing a portion of a rate structure, defined in the
\href{https://github.com/California-Data-Collaborative/Open-Water-Rate-Specification}{OWRS file}}
}
\value{
Either a single value or a vector the length of \code{df}. The result of
evaluating a field or formula in the context of the dataframe.
}
\description{
\code{eval_field_or_formula} evaluates a field or formula in the context of
the provided dataframe.
}
\keyword{internal}

