% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateParser.R
\name{eval_map}
\alias{eval_map}
\title{Evaluate a map.}
\usage{
eval_map(df, rate_part)
}
\arguments{
\item{df}{Data frame containing all data values referenced in the map.}

\item{rate_part}{List representing a portion of a rate structure, defined in the
\href{https://github.com/California-Data-Collaborative/Open-Water-Rate-Specification}{OWRS file}}
}
\value{
Vector the length of \code{df}. The result of evaluating a map in the context
of the dataframe.
}
\description{
\code{eval_map} evaluates a conditional/map in the context of
the provided dataframe.
}
\keyword{internal}

