% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateParser.R
\name{is_map}
\alias{is_map}
\title{Check whether a rate part is a map.}
\usage{
is_map(rate_part)
}
\arguments{
\item{rate_part}{List representing a portion of a rate structure, defined in the
\href{https://github.com/California-Data-Collaborative/Open-Water-Rate-Specification}{OWRS file}}
}
\value{
Boolean representing whether the rate part is a map.
}
\description{
\code{is_map} checks whether the rate part contains a \code{depends_on}
and a \code{values} clause. If it does, then it is a proper map.
}

