% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rateParser.R
\name{read_owrs_file}
\alias{read_owrs_file}
\title{Read an OWRS file.}
\usage{
read_owrs_file(filepath)
}
\arguments{
\item{filepath}{Path to the OWRS file.}
}
\value{
Nested list representing the parsed yaml structure.
}
\description{
\code{read_owrs_file} reads a raw \code{yaml} file and parses it
into a list structure.
}

